remove unneded usages, moved helper to proper directory
