+++ src/Sylius/Behat/Resources/config/services.xml
@@ -88,7 +88,9 @@
 
         <service id="sylius.behat.shared_storage" class="Sylius\Behat\Service\SharedStorage" public="false"/>
 
-        <service id="sylius.behat.java_script_test_helper" class="Sylius\Behat\JavaScriptTestHelper" />
+        <service id="sylius.behat.java_script_test_helper" class="Sylius\Behat\Service\Helper\JavaScriptTestHelper" >
+            <argument>100000</argument>
+        </service>
 
         <service id="sylius.calendar" class="Sylius\Behat\Service\Provider\Calendar">
             <argument>%kernel.project_dir%</argument>
